+++ R/lfa.R
@@ -8,18 +8,24 @@
 #'
 #' @param X A matrix of SNP genotypes, i.e. an integer matrix of 0's,
 #' 1's, 2's and `NA`s.
+#' BEDMatrix is supported.
 #' Sparse matrices of class Matrix are not supported (yet).
 #' @param d Number of logistic factors, including the intercept
 #' @param adjustments A matrix of adjustment variables to hold fixed during estimation.
 #' Number of rows must equal number of individuals in `X`.
 #' These adjustments take the place of LFs in the output, so the number of columns must not exceed `d-2` to allow for the intercept and at least one proper LF to be included.
 #' When present, these adjustment variables appear in the first columns of the output.
+#' Not supported when `X` is a BEDMatrix object.
 #' @param override Optional boolean to bypass Lanczos bidiagonalization
 #' SVD. Usually not advised unless encountering a bug in the SVD code.
+#' Ignored if `X` is a BEDMatrix object.
 #' @param safety Optional boolean to bypass checks on the genotype
 #' matrices, which require a non-trivial amount of computation.
+#' Ignored if `X` is a BEDMatrix object.
 #' @param ploidy Ploidy of data, defaults to 2 for bi-allelic unphased SNPs
 #' @param tol Tolerance value passed to `trunc_svd`
+#' Ignored if `X` is a BEDMatrix object.
+#' @param m_chunk If `X` is a BEDMatrix object, number of loci to read per chunk (to control memory usage).
 #' 
 #' @return The matrix of logistic factors, with individuals along rows and factors along columns.
 #' The intercept appears at the end of the columns, and adjustments in the beginning if present.
@@ -38,20 +44,30 @@ lfa <- function(
                 override = FALSE,
                 safety = FALSE,
                 ploidy = 2,
-                tol = 1e-13
+                tol = 1e-13,
+                m_chunk = 1000 # gave good performance in tests
                 ) {
     if ( missing( X ) )
         stop( 'Genotype matrix `X` is required!' )
     if ( missing( d ) )
-        stop( 'Dimension number `d` is missing!' )
-    
-    # check data if asked to do it
-    if ( safety )
-        check_geno(X)
+        stop( 'Dimension number `d` is required!' )
 
+    # check class
+    is_BEDMatrix <- FALSE
+    if ( "BEDMatrix" %in% class(X) ) {
+        is_BEDMatrix <- TRUE
+    } else if ( !is.matrix( X ) )
+        stop( '`X` must be a matrix!' )
+    
     # data dimensions
-    m <- nrow(X)
-    n <- ncol(X)
+    if ( is_BEDMatrix ) {
+        # matrix is transposed in this case
+        # only this one is needed until the pure BEDMatrix code kicks in
+        n <- nrow(X)
+    } else {
+        m <- nrow(X)
+        n <- ncol(X)
+    }
 
     # check for d validity
     if( !is.numeric(d) ) {
@@ -66,19 +82,29 @@ lfa <- function(
     } else if( d > 1 ) {
         d <- d - 1 #for the svd stuff
     }
-    
+
     # check adjustments vars
     if ( !is.null( adjustments ) ){
+        if ( is_BEDMatrix )
+            stop( '`adjustments` are not supported when `X` is a BEDMatrix object!' )
         if ( !is.matrix( adjustments ) )
             stop( '`adjustments` must be a matrix!' )
-        if ( nrow( adjustments ) != ncol( X ) )
+        if ( nrow( adjustments ) != n )
             stop( "`adjustments` needs to have same number of rows as individuals" )
         if ( ncol( adjustments ) >= d )
             stop("need to estimate at least one non-adjustment logistic factor")
-        if ( sum( ! stats::complete.cases( adjustments ) ) > 0 )
-            stop( "no missing values in adjustments" )
+        if ( anyNA( adjustments ) )
+            stop( '`adjustments` must not have missing values!' )
     }
 
+    if ( is_BEDMatrix )
+        return( lfa_BEDMatrix( X, d, ploidy = ploidy, m_chunk = m_chunk ) )
+    # else continue
+    
+    # check data if asked to do it
+    if ( safety )
+        check_geno(X)
+    
     # a mysterious parameter for trunc_svd
     adjust <- 8
     if ( n - d < 10 )